+++ water_neuron/src/main.rs
@@ -1,10 +1,9 @@
-use candid::{Decode, Nat, Principal};
+use candid::{Nat, Principal};
 use ic_canister_log::log;
 use ic_canisters_http_types::{HttpRequest, HttpResponse, HttpResponseBuilder};
 use ic_cdk_macros::{init, post_upgrade, query, update};
 use ic_metrics_encoder::MetricsEncoder;
 use icrc_ledger_types::icrc1::account::Account;
-use strum::IntoEnumIterator;
 use water_neuron::conversion::{MINIMUM_DEPOSIT_AMOUNT, MINIMUM_WITHDRAWAL_AMOUNT};
 use water_neuron::dashboard::DisplayAmount;
 use water_neuron::guards::GuardPrincipal;
@@ -23,11 +22,11 @@ use water_neuron::state::{
 use water_neuron::storage::total_event_count;
 use water_neuron::tasks::{schedule_now, TaskType};
 use water_neuron::{
-    convert_amount_e8s_to_string_representation, CancelWithdrawalError, CanisterInfo, ConsentInfo,
-    ConsentMessage, ConsentMessageMetadata, ConsentMessageRequest, ConversionArg, ConversionError,
-    DepositSuccess, ErrorInfo, Icrc21Error, Icrc21Function, LiquidArg, StandardRecord, Unit,
-    UpgradeArg, WithdrawalSuccess, MAX_CONSENT_MESSAGE_ARG_SIZE_BYTES,
+    CancelWithdrawalError, CanisterInfo, ConversionArg, ConversionError, DepositSuccess, LiquidArg,
+    Unit, UpgradeArg, WithdrawalSuccess,
 };
+
+pub mod icrc21;
 
 fn reject_anonymous_call() {
     if ic_cdk::caller() == Principal::anonymous() {
@@ -170,86 +169,6 @@ fn get_wtn_proposal_id(nns_proposal_id: u64) -> Result<ProposalId, ProposalId> {
             })
             .cloned()
             .ok_or_else(|| s.last_nns_proposal_seen.clone())
-    })
-}
-
-#[query]
-fn icrc10_supported_standards() -> Vec<StandardRecord> {
-    vec![
-        StandardRecord {
-            name: "ICRC-21".to_string(),
-            url: "https://github.com/dfinity/wg-identity-authentication/blob/main/topics/ICRC-21/icrc_21_consent_msg.md".to_string(),
-        }
-    ]
-}
-
-#[update]
-fn icrc21_canister_call_consent_message(
-    request: ConsentMessageRequest,
-) -> Result<ConsentInfo, Icrc21Error> {
-    if request.arg.len() > MAX_CONSENT_MESSAGE_ARG_SIZE_BYTES as usize {
-        return Err(Icrc21Error::UnsupportedCanisterCall(ErrorInfo {
-            description: format!(
-                "The argument size is too large. The maximum allowed size is {} bytes.",
-                MAX_CONSENT_MESSAGE_ARG_SIZE_BYTES
-            ),
-        }));
-    }
-
-    let metadata = ConsentMessageMetadata {
-        language: "en".to_string(),
-        utc_offset_minutes: request.user_preferences.metadata.utc_offset_minutes,
-    };
-
-    let message = match request.method.parse::<Icrc21Function>().map_err(|err| {
-        Icrc21Error::UnsupportedCanisterCall(ErrorInfo {
-            description: format!(
-                "The call provided is not supported: {}.\n Supported calls for ICRC-21 are {:?}. \n Error: {:?}", 
-                request.method,
-                Icrc21Function::iter().map(|f|f.to_string()).collect::<Vec<String>>(),
-                err
-            ),
-        })
-    })? {
-        Icrc21Function::Stake =>  {
-            let arg = Decode!(&request.arg, ConversionArg).map_err(|e| Icrc21Error::UnsupportedCanisterCall(ErrorInfo {
-                description: format!("Failed to decode ConversionArg: {}", e)
-            }))?;
-            let icp_amount = convert_amount_e8s_to_string_representation(arg.amount_e8s, 8);
-            match arg.maybe_subaccount {
-                Some(subaccount) => format!("Convert {icp_amount} ICP to nICP at the current exchange rate. Specified subaccount: {subaccount:?}."),
-                None => format!("Convert {icp_amount} ICP to nICP at the current exchange rate.")
-            }
-        },
-        Icrc21Function::Unstake =>  {
-            let arg = Decode!(&request.arg, ConversionArg).map_err(|e| Icrc21Error::UnsupportedCanisterCall(ErrorInfo {
-                description: format!("Failed to decode ConversionArg: {}", e),
-            }))?;
-            let nicp_amount = convert_amount_e8s_to_string_representation(arg.amount_e8s, 8);
-            match arg.maybe_subaccount {
-                Some(subaccount) => format!(
-                    "Convert {nicp_amount} nICP to ICP at the current exchange rate after a 6 months dissolve delay. 
-                    Specified subaccount: {subaccount:?}."
-                ),
-                None => format!(
-                    "Convert {nicp_amount} nICP to ICP at the current exchange rate after a 6 months dissolve delay."
-                )
-            }
-        },
-        Icrc21Function::CancelWithdrawal =>  {
-            let arg = Decode!(&request.arg, NeuronId).map_err(|e| Icrc21Error::UnsupportedCanisterCall(ErrorInfo {
-                description: format!("Failed to decode NeuronId: {}", e),
-            }))?;
-            format!("Cancel the withdrawal associated to the neuron id: {}.", arg.id)
-        },
-        Icrc21Function::ClaimAirdrop =>  {
-            format!("Claim WTN tokens associated to your airdrop allocation. ")
-        }
-    };
-
-    Ok(ConsentInfo {
-        metadata,
-        consent_message: ConsentMessage::GenericDisplayMessage(message),
     })
 }
